+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- Fix: Add dynamic resizing for the charts on Stats pages
 - Fix: Instructor's hours for the Total Stats table were mixed up for the last 90 days and the last 12 months
 
 ## [2.35.1] - 20.05.2024